Asset refactoring, path loading, and debugging
Asset now keeps an "assetPath" property so assets can be loaded from
URLs. This has been used so that sprites, maps, and audio resources no
longer are loaded statically from the "assets" folder, but may be
loaded

Asset config files now have an optional debug attribute, which is an
object containing debug parameters. Currently the only attribute
supported is "drawCollision," which draws collision data for sprites.

Actor animation assets have been moved from "actors/strips" to the
"animations" folder, and code has been updated for this change.

Specified "build" as the files to use when this is installed as a
dependency.

Removed superfluous and outdated test assets.