Packaging refactoring
Launcher refactored so that the game will launch when "launcher" is the
main process, which it has been doing up til now, but if not it will not
launch immediately. This way yage can become a dependency of other node
projects, and then you can call yage('path/to/assets/') to launch the
game with that assets folder.