patch 0.4.1 - behavior system
Quick fix to behavior here: The first arguments go in first, but the
return is actually a function that takes the state parameter. Or, to say
it differently, state is now curried. This is important for making
things easier with passing state with _.pipe/_.flow functions in
different branches where I test things with behavior