Actor preloading
A map keeps a record of animations used by the actors in sprites when it
is initialized. This record is used to preload the images for these
animations, so that when animations change there isn't the visible
flicker that occurs otherwise (This is obvious when you switch george to
george spinning. However, in previous commits this might not appear
since this feature was tested previously by having the image preloaded
through direct inclusion in main-window.html)