Master sprites
Sprites can now be added to the sprite list in config files without
physics options/empty physics. When this is done, they are listed
as "master sprites," and removed from the sprite list. This is so these
sprites can be cloned later for dynamically created sprites, while still
letting us know what sprites will be in a map.

Sprites like this can also be included with a single string if they
are a template sprite.