Animation offset change
Animation offset changed: From now on, it will be used only when an
animation is set or changed, and it will update the actual coordinates
of the sprite. This means that the posX and posY in the game state will
represent the x and y coordinates of the bounding box of the sprite.
This will make it easier for physics since we don't have to recalculate
the "real" bounding box each time from the offset.